+++ src/app/private-event/private-event.component.ts
@@ -10,7 +10,7 @@ import {
   IonLabel,
   IonInput,
   IonModal,
-  IonPicker,
+  IonPickerLegacy,
   IonTitle,
   IonToolbar,
   ModalController,
@@ -42,7 +42,7 @@ export enum PrivateEventResult {
     IonModal,
     IonDatetimeButton,
     IonDatetime,
-    IonPicker,
+    IonPickerLegacy,
     IonContent,
     IonInput,
     IonButtons,
@@ -64,13 +64,13 @@ export class PrivateEventComponent implements OnInit {
   public startEvent = new Date(new Date().getFullYear(), 7, 20).toISOString();
   public endEvent = new Date(new Date().getFullYear(), 8, 10).toISOString();
 
-  event = model<PrivateEvent>({
+  event: PrivateEvent = {
     title: '',
     id: uniqueId('pe'),
     start: this.initialTime,
     address: this.noAddress,
     notes: '',
-  });
+  };
   constructor() {
     this.addresses = this.streetService.getAddresses();
   }
@@ -83,13 +83,13 @@ export class PrivateEventComponent implements OnInit {
     {
       text: 'Confirm',
       handler: (value: any) => {
-        this.event().address = `${value.hour.value}:${value.minute.value} & ${value.street.value}`;
+        this.event.address = `${value.hour.value}:${value.minute.value} & ${value.street.value}`;
       },
     },
   ];
 
   ngOnInit() {
-    this.streetService.setAddress(this.event().address, this.addresses);
+    this.streetService.setAddress(this.event.address, this.addresses);
   }
 
   cancel() {
@@ -97,23 +97,23 @@ export class PrivateEventComponent implements OnInit {
   }
 
   confirm() {
-    if (this.event().title.length == 0) {
+    if (this.event.title.length == 0) {
       this.presentToast(`Specify name`);
       return;
     }
-    if (this.event().address == this.noAddress) {
+    if (this.event.address == this.noAddress) {
       this.presentToast(`Select an address`);
       return;
     }
-    if (this.event().start == this.initialTime) {
+    if (this.event.start == this.initialTime) {
       this.presentToast(`Select a date and time when the event starts`);
       return;
     }
-    return this.modalCtrl.dismiss(this.event(), PrivateEventResult.confirm);
+    return this.modalCtrl.dismiss(this.event, PrivateEventResult.confirm);
   }
 
   deleteEvent() {
-    return this.modalCtrl.dismiss(this.event(), PrivateEventResult.delete);
+    return this.modalCtrl.dismiss(this.event, PrivateEventResult.delete);
   }
 
   async presentToast(message: string) {