+++ changelog.md
@@ -5,6 +5,7 @@
 - Print option for favorites
 - Fix for error when trying to download maps that are > 5mb
 - Move event images to bottom of popover so you can access information
+- Fix for private event editing which was previously blank
 
 ## 2.14